% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paccpack-package.R
\docType{data}
\name{characteristics_of_counties}
\alias{characteristics_of_counties}
\title{Characteristics of Iranian's counties.}
\format{
A data frame with 418 rows and 15 variables:
\describe{
\item{state}{names of states as factor}
\item{county}{names of counties as factor}
\item{area}{area of each county in square meters}
\item{pop_2016}{population of each county in 2016}
\item{pop_2006}{population of each county in 2006}
\item{pop_1996}{population of each county in 1996}
\item{emp_2016}{employment of each county in 2016}
\item{emp_2006}{employment of each county in 2006}
\item{emp_1996}{employment of each county in 1996}
\item{tii}{internal travel time within each county in second}
\item{tij_2016}{the shortest travel time between the county i and county j in second- 2016}
\item{tij_2006}{the shortest travel time between the county i and county j in second- 2006}
\item{tij_1996}{the shortest travel time between the county i and county j in second- 1996}
\item{R_2016_2006}{employment growth rate of each county - 2006_2016}
\item{R_2006_1996}{employment growth rate of each county - 1996_2016}
}
}
\source{
statistical Center of Iran (1996, 2006 ,2016) General census of population and housing of Iran (https://www.amar.org.ir) and The Railway Organization and Road Maintenance and Transportation Organazation of Iran (1996, 2006 ,2016)
}
\usage{
characteristics_of_counties
}
\description{
Characteristics of Iranian's counties
}
\details{
A dataset containing information about 418 counties in Iran, including
population and employment estimates, Employment rate,area,and travel time.
}
\keyword{datasets}
