% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paccpack-package.R
\docType{data}
\name{potential_accessibility}
\alias{potential_accessibility}
\title{Potential accessibility
A dataset containing information about potential accessibility of 418 counties in Iran.}
\format{
A data frame with 1254 rows and 6 variables:
Accessibility to transportation network based on population or employment
\describe{
\item{population}{Population of each county in 1996, 2006 and 2016}
\item{employment}{Employment of each county in 1996, 2006 and 2016}
\item{year}{year of the estimates}
\item{pacc_p}{potential of Accessibility based on population in 1996, 2006 and 2016}
\item{pacc_e}{potential of Accessibility based on employment in 1996, 2006 and 2016}
}
}
\source{
Population and employment Data Source: statistical Center of Iran (1996, 2006 ,2016) General census of population and housing of Iran (https://www.amar.org.ir)
}
\usage{
potential_accessibility
}
\description{
potential accessibility of Iranian's counties based on population and employment.
}
\keyword{datasets}
